requirements.txt
alembic==1.7.4
aniso8601==9.0.1
attrs==21.2.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
dnspython==2.1.0
docx==0.2.4
eventlet==0.32.0
Flask==1.1.2
Flask-Migrate==2.5.2
flask-redis==0.4.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonpickle==1.1
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.0.0
packaging==21.2
passlib==1.7.4
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-docx==0.8.11
python-engineio==4.3.0
python-socketio==5.4.1
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
